refactor: rename Verbs to more precise RBACVerbs
